+++ src/api_insee/utils/client_credentials.py
@@ -1,12 +1,10 @@
 import base64
 
-from typing import Optional
-
 from api_insee.exeptions.authentication_error import InvalidCredentialsError
 
 
 class ClientCredentials:
-    def __init__(self, key: Optional[str] = None, secret: Optional[str] = None) -> None:
+    def __init__(self, key: str, secret: str) -> None:
         self.key = key
         self.secret = secret