+++ tests/conftest.py
@@ -1,26 +1,12 @@
 import re
 
-from pathlib import Path
-from typing import Any, Dict, Final
+from typing import Any, Dict, Type
 
 import pytest
 
 from _pytest.fixtures import SubRequest
 from api_insee import ApiInsee
-
-PROJECT_ROOT: Final = Path(__file__).parent.parent
-
-
-def parse_env_file() -> Dict[str, str]:
-    with Path(PROJECT_ROOT, ".env").open() as file:
-        lines = (
-            line.split("=")
-            for line in file.read().splitlines(keepends=False)
-            if line and not line.startswith("#")
-        )
-
-        return {key.strip(): value.strip() for key, value in lines}
-
+from utils import ApiInseeTestDouble, parse_env_file
 
 CREDENTIALS = parse_env_file()
 SIRENE_API_CONSUMER_KEY = CREDENTIALS["SIRENE_API_CONSUMER_KEY"]
@@ -29,11 +15,24 @@ SIRENE_API_CONSUMER_SECRET = CREDENTIALS["SIRENE_API_CONSUMER_SECRET"]
 
 @pytest.fixture()
 def api(request: SubRequest) -> ApiInsee:
-    return ApiInsee(
+    api_insee: Type[ApiInsee] = (
+        ApiInsee
+        if request.config.option.record_mode is not None
+        else ApiInseeTestDouble
+    )
+
+    return api_insee(
         SIRENE_API_CONSUMER_KEY,
         SIRENE_API_CONSUMER_SECRET,
-        noauth=request.config.option.record_mode is None,
     )
+
+
+@pytest.fixture()
+def vcr_config() -> Dict[str, Any]:
+    return {
+        "filter_headers": ["authorization", "api_token", "Set-Cookie"],
+        "before_record_response": replace_token,
+    }
 
 
 def replace_token(response: Dict[str, Any]) -> Dict[str, Any]:
@@ -48,11 +47,3 @@ def replace_token(response: Dict[str, Any]) -> Dict[str, Any]:
         )
 
     return response
-
-
-@pytest.fixture()
-def vcr_config() -> Dict[str, Any]:
-    return {
-        "filter_headers": ["authorization", "api_token", "Set-Cookie"],
-        "before_record_response": replace_token,
-    }